feat:joins and types
This PR adds the ability to specify a foreign key for a dbt model in dal. The generated GraphQL api then supports traversing that relationship.

It works via the standard dataloader approach, which has some obvious downsides for column oriented data warehouses. It's simple though and doesn't try to fight GraphQL. It also makes it really easy to implement entity caching.

The PR also adds a very basic v1 type system to dal.